Implement most of the classes in the bean package Fix the .iml project file

+++ src/main/java/lt/dndfan/dndhelper/bean/character/inventory/ICharacterInventory.kt
@@ -8,5 +8,5 @@ interface ICharacterInventory : Iterable<IItem> {
     fun addItem(item : IItem)
     fun removeItem(item : IItem)
 
-    fun getWeight() : Int
+    fun getWeight() : Double
 }
